control-box-function: remove conversion warning

+++ include/proxddp/modelling/control-box-function.hxx
@@ -7,7 +7,8 @@ template <typename Scalar>
 ControlBoxFunctionTpl<Scalar>::ControlBoxFunctionTpl(const int ndx,
                                                      const VectorXs umin,
                                                      const VectorXs umax)
-    : Base(ndx, umin.size(), ndx, 2 * umin.size()), umin_(umin), umax_(umax) {
+    : Base(ndx, (int)umin.size(), ndx, 2 * (int)umin.size()), umin_(umin),
+      umax_(umax) {
   assert(umin.size() == umax.size() &&
          "Size of umin and umax should be the same!");
 }
